Fix some parsing issues

+++ web-ifc-cpp/util.h
@@ -47,9 +47,9 @@ namespace webifc
 			points.push_back(c);
 
 			Face f;
-			f.i0 = points.size() - 3;
-			f.i1 = points.size() - 2;
-			f.i2 = points.size() - 1;
+			f.i0 = static_cast<uint32_t>(points.size() - 3);
+			f.i1 = static_cast<uint32_t>(points.size() - 2);
+			f.i2 = static_cast<uint32_t>(points.size() - 1);
 
 			faces.push_back(f);
 		}
